requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
kaggle>=1.6.0
